# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reverie API",
    "description": "HTTP API for Reverie, a self-hosted ebook library manager. Generated from the server handlers; do not edit by hand.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/dashboard/activity": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "`GET /api/v1/dashboard/activity?limit=N` — most-recent ingestion batches,\nnewest first (admin only).",
        "description": "Groups `ingestion_jobs` by `batch_id` into per-batch outcome counts.\n`acquire_with_rls` is used for pattern consistency but is not load-bearing\nhere — `ingestion_jobs` has no RLS.\n\n# Errors\n- [`AppError::Forbidden`] when the caller is not an admin.\n- [`AppError::Internal`] on database errors.",
        "operationId": "activity",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Max recent batches to return; clamped to 1..=100, default 20.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Most-recent ingestion batches, newest first. Admin only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/stats": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "`GET /api/v1/dashboard/stats` — library-wide aggregate health metrics\n(admin only).",
        "description": "Computes the manifestation-side metrics in a single conditional-aggregate\nscan joined to `works` (query A), plus one per-format `GROUP BY` (query\nB). Validation/enrichment breakdowns are reassembled in Rust from the\nfixed `FILTER` columns so the closed enum sets render deterministically.\n\n# Errors\n- [`AppError::Forbidden`] when the caller is not an admin.\n- [`AppError::Internal`] on database errors.",
        "operationId": "stats",
        "responses": {
          "200": {
            "description": "Library-wide aggregate health metrics. Admin only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityResponse": {
        "type": "object",
        "description": "Response shape for `GET /api/v1/dashboard/activity`.",
        "required": [
          "batches"
        ],
        "properties": {
          "batches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchRow"
            }
          }
        }
      },
...
      "BatchRow": {
        "type": "object",
        "description": "One recent ingestion batch. Invariant:\n`completed + failed + skipped + in_progress == total`. `ended_at` is null\nwhile the batch is still in flight (any `queued`/`running` job).",
        "required": [
          "batch_id",
          "started_at",
          "total",
          "completed",
          "failed",
          "skipped",
          "in_progress"
        ],
        "properties": {
          "batch_id": {
            "type": "string",
            "format": "uuid"
          },
          "completed": {
            "type": "integer",
            "format": "int64"
          },
          "ended_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "failed": {
            "type": "integer",
            "format": "int64"
          },
          "in_progress": {
            "type": "integer",
            "format": "int64"
          },
          "skipped": {
            "type": "integer",
            "format": "int64"
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "FormatBucket": {
        "type": "object",
        "description": "One per-format storage bucket (`GROUP BY format`).",
        "required": [
          "format",
          "count",
          "bytes"
        ],
        "properties": {
          "bytes": {
            "type": "integer",
            "format": "int64"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "format": {
            "type": "string"
          }
        }
      },
...
      "MetadataCoverage": {
        "type": "object",
        "description": "Metadata-coverage numerators (denominator is `total`). The frontend\nderives percentages so the wire stays integer-only.",
        "required": [
          "total",
          "has_description",
          "has_language",
          "has_isbn_13",
          "has_cover"
        ],
        "properties": {
          "has_cover": {
            "type": "integer",
            "format": "int64"
          },
          "has_description": {
            "type": "integer",
            "format": "int64"
          },
          "has_isbn_13": {
            "type": "integer",
            "format": "int64"
          },
          "has_language": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 `application/problem+json` error body — the single runtime DTO\nfor the error envelope. The `IntoResponse` impl on\n[`crate::error::AppError`] constructs it for every Problem-Details\nvariant, and operational probes (readiness) build it directly, so the\ndocumented component schema and the bytes on the wire come from the same\nstruct and cannot drift.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "URI reference identifying the specific request, when available.\nOmitted (RFC 9457 §3.1 permits this) outside an HTTP request."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code, repeated in the body per RFC 9457.",
            "example": 404,
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem type."
          },
          "type": {
            "type": "string",
            "description": "Stable URI reference identifying the problem type.",
            "example": "https://reverie.example/probs/not-found"
          }
        }
      },
...
      "StatsResponse": {
        "type": "object",
        "description": "Response shape for `GET /api/v1/dashboard/stats`.",
        "required": [
          "total_manifestations",
          "total_works",
          "storage_total_bytes",
          "storage_cover_bytes",
          "storage_by_format",
          "validation_breakdown",
          "clean_non_epub_count",
          "enrichment_breakdown",
          "metadata_coverage"
        ],
        "properties": {
          "clean_non_epub_count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of `clean` rows that are non-EPUB formats. UNK-313: non-EPUB\nfiles carry `validation_status='clean'` despite never being\nstructurally validated — surfaced so the UI can footnote the bucket."
          },
          "enrichment_breakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatusCount"
            }
          },
          "metadata_coverage": {
            "$ref": "#/components/schemas/MetadataCoverage"
          },
          "storage_by_format": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FormatBucket"
            }
          },
          "storage_cover_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "storage_total_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "total_manifestations": {
            "type": "integer",
            "format": "int64"
          },
          "total_works": {
            "type": "integer",
            "format": "int64"
          },
          "validation_breakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatusCount"
            }
          }
        }
      },
      "StatusCount": {
        "type": "object",
        "description": "One `{ status, count }` bucket in a breakdown array. `status` is the\ncanonical wire label of a [`ValidationStatus`] / [`EnrichmentStatus`]\nvariant, so the closed enum set is reflected on the wire even for zero\ncounts.",
        "required": [
          "status",
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "opds_basic": {
        "type": "http",
        "scheme": "basic",
        "description": "HTTP Basic authentication for the OPDS feeds. MUST be used over HTTPS in production — Basic credentials are otherwise exposed in transit (Checkov CKV_OPENAPI_3)."
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "id",
        "description": "Session cookie issued by the server's session layer. Set with the Secure attribute when `behind_https` is enabled; MUST be served over HTTPS in production to prevent session hijacking."
      }
    }
  },
  "security": [
    {
      "session_cookie": []
    }
  ],
...
}